@@ -1,4 +1,4 @@
-// +build linux
+//go:build linux
 
 /*
  * Tencent is pleased to support the open source community by making TKEStack available.
