@@ -1,3 +1,5 @@
+//go:build linux
+
 /*
  * Tencent is pleased to support the open source community by making TKEStack available.
  *
@@ -14,7 +16,6 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-// +build linux
 
 package main
 
