@@ -1,24 +1,26 @@
 /*
-Copyright The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
+ * Tencent is pleased to support the open source community by making TKEStack available.
+ *
+ * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at
+ *
+ * https://opensource.org/licenses/Apache-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 // Code generated by client-gen. DO NOT EDIT.
 
 package fake
 
 import (
+	"context"
+
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
@@ -38,7 +40,7 @@ var floatingipsResource = schema.GroupVersionResource{Group: "galaxy.k8s.io", Ve
 var floatingipsKind = schema.GroupVersionKind{Group: "galaxy.k8s.io", Version: "v1alpha1", Kind: "FloatingIP"}
 
 // Get takes name of the floatingIP, and returns the corresponding floatingIP object, and an error if there is any.
-func (c *FakeFloatingIPs) Get(name string, options v1.GetOptions) (result *v1alpha1.FloatingIP, err error) {
+func (c *FakeFloatingIPs) Get(ctx context.Context, name string, options v1.GetOptions) (result *v1alpha1.FloatingIP, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewRootGetAction(floatingipsResource, name), &v1alpha1.FloatingIP{})
 	if obj == nil {
@@ -48,7 +50,7 @@ func (c *FakeFloatingIPs) Get(name string, options v1.GetOptions) (result *v1alp
 }
 
 // List takes label and field selectors, and returns the list of FloatingIPs that match those selectors.
-func (c *FakeFloatingIPs) List(opts v1.ListOptions) (result *v1alpha1.FloatingIPList, err error) {
+func (c *FakeFloatingIPs) List(ctx context.Context, opts v1.ListOptions) (result *v1alpha1.FloatingIPList, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewRootListAction(floatingipsResource, floatingipsKind, opts), &v1alpha1.FloatingIPList{})
 	if obj == nil {
@@ -69,13 +71,13 @@ func (c *FakeFloatingIPs) List(opts v1.ListOptions) (result *v1alpha1.FloatingIP
 }
 
 // Watch returns a watch.Interface that watches the requested floatingIPs.
-func (c *FakeFloatingIPs) Watch(opts v1.ListOptions) (watch.Interface, error) {
+func (c *FakeFloatingIPs) Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error) {
 	return c.Fake.
 		InvokesWatch(testing.NewRootWatchAction(floatingipsResource, opts))
 }
 
 // Create takes the representation of a floatingIP and creates it.  Returns the server's representation of the floatingIP, and an error, if there is any.
-func (c *FakeFloatingIPs) Create(floatingIP *v1alpha1.FloatingIP) (result *v1alpha1.FloatingIP, err error) {
+func (c *FakeFloatingIPs) Create(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.CreateOptions) (result *v1alpha1.FloatingIP, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewRootCreateAction(floatingipsResource, floatingIP), &v1alpha1.FloatingIP{})
 	if obj == nil {
@@ -85,7 +87,7 @@ func (c *FakeFloatingIPs) Create(floatingIP *v1alpha1.FloatingIP) (result *v1alp
 }
 
 // Update takes the representation of a floatingIP and updates it. Returns the server's representation of the floatingIP, and an error, if there is any.
-func (c *FakeFloatingIPs) Update(floatingIP *v1alpha1.FloatingIP) (result *v1alpha1.FloatingIP, err error) {
+func (c *FakeFloatingIPs) Update(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.UpdateOptions) (result *v1alpha1.FloatingIP, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewRootUpdateAction(floatingipsResource, floatingIP), &v1alpha1.FloatingIP{})
 	if obj == nil {
@@ -95,22 +97,22 @@ func (c *FakeFloatingIPs) Update(floatingIP *v1alpha1.FloatingIP) (result *v1alp
 }
 
 // Delete takes name of the floatingIP and deletes it. Returns an error if one occurs.
-func (c *FakeFloatingIPs) Delete(name string, options *v1.DeleteOptions) error {
+func (c *FakeFloatingIPs) Delete(ctx context.Context, name string, opts v1.DeleteOptions) error {
 	_, err := c.Fake.
-		Invokes(testing.NewRootDeleteAction(floatingipsResource, name), &v1alpha1.FloatingIP{})
+		Invokes(testing.NewRootDeleteActionWithOptions(floatingipsResource, name, opts), &v1alpha1.FloatingIP{})
 	return err
 }
 
 // DeleteCollection deletes a collection of objects.
-func (c *FakeFloatingIPs) DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error {
-	action := testing.NewRootDeleteCollectionAction(floatingipsResource, listOptions)
+func (c *FakeFloatingIPs) DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error {
+	action := testing.NewRootDeleteCollectionAction(floatingipsResource, listOpts)
 
 	_, err := c.Fake.Invokes(action, &v1alpha1.FloatingIPList{})
 	return err
 }
 
 // Patch applies the patch and returns the patched floatingIP.
-func (c *FakeFloatingIPs) Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.FloatingIP, err error) {
+func (c *FakeFloatingIPs) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.FloatingIP, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewRootPatchSubresourceAction(floatingipsResource, name, pt, data, subresources...), &v1alpha1.FloatingIP{})
 	if obj == nil {
