@@ -1,24 +1,26 @@
 /*
-Copyright The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
+ * Tencent is pleased to support the open source community by making TKEStack available.
+ *
+ * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at
+ *
+ * https://opensource.org/licenses/Apache-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 // Code generated by client-gen. DO NOT EDIT.
 
 package fake
 
 import (
+	"context"
+
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
@@ -39,7 +41,7 @@ var poolsResource = schema.GroupVersionResource{Group: "galaxy.k8s.io", Version:
 var poolsKind = schema.GroupVersionKind{Group: "galaxy.k8s.io", Version: "v1alpha1", Kind: "Pool"}
 
 // Get takes name of the pool, and returns the corresponding pool object, and an error if there is any.
-func (c *FakePools) Get(name string, options v1.GetOptions) (result *v1alpha1.Pool, err error) {
+func (c *FakePools) Get(ctx context.Context, name string, options v1.GetOptions) (result *v1alpha1.Pool, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewGetAction(poolsResource, c.ns, name), &v1alpha1.Pool{})
 
@@ -50,7 +52,7 @@ func (c *FakePools) Get(name string, options v1.GetOptions) (result *v1alpha1.Po
 }
 
 // List takes label and field selectors, and returns the list of Pools that match those selectors.
-func (c *FakePools) List(opts v1.ListOptions) (result *v1alpha1.PoolList, err error) {
+func (c *FakePools) List(ctx context.Context, opts v1.ListOptions) (result *v1alpha1.PoolList, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewListAction(poolsResource, poolsKind, c.ns, opts), &v1alpha1.PoolList{})
 
@@ -72,14 +74,14 @@ func (c *FakePools) List(opts v1.ListOptions) (result *v1alpha1.PoolList, err er
 }
 
 // Watch returns a watch.Interface that watches the requested pools.
-func (c *FakePools) Watch(opts v1.ListOptions) (watch.Interface, error) {
+func (c *FakePools) Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error) {
 	return c.Fake.
 		InvokesWatch(testing.NewWatchAction(poolsResource, c.ns, opts))
 
 }
 
 // Create takes the representation of a pool and creates it.  Returns the server's representation of the pool, and an error, if there is any.
-func (c *FakePools) Create(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err error) {
+func (c *FakePools) Create(ctx context.Context, pool *v1alpha1.Pool, opts v1.CreateOptions) (result *v1alpha1.Pool, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewCreateAction(poolsResource, c.ns, pool), &v1alpha1.Pool{})
 
@@ -90,7 +92,7 @@ func (c *FakePools) Create(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err erro
 }
 
 // Update takes the representation of a pool and updates it. Returns the server's representation of the pool, and an error, if there is any.
-func (c *FakePools) Update(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err error) {
+func (c *FakePools) Update(ctx context.Context, pool *v1alpha1.Pool, opts v1.UpdateOptions) (result *v1alpha1.Pool, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewUpdateAction(poolsResource, c.ns, pool), &v1alpha1.Pool{})
 
@@ -101,23 +103,23 @@ func (c *FakePools) Update(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err erro
 }
 
 // Delete takes name of the pool and deletes it. Returns an error if one occurs.
-func (c *FakePools) Delete(name string, options *v1.DeleteOptions) error {
+func (c *FakePools) Delete(ctx context.Context, name string, opts v1.DeleteOptions) error {
 	_, err := c.Fake.
-		Invokes(testing.NewDeleteAction(poolsResource, c.ns, name), &v1alpha1.Pool{})
+		Invokes(testing.NewDeleteActionWithOptions(poolsResource, c.ns, name, opts), &v1alpha1.Pool{})
 
 	return err
 }
 
 // DeleteCollection deletes a collection of objects.
-func (c *FakePools) DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error {
-	action := testing.NewDeleteCollectionAction(poolsResource, c.ns, listOptions)
+func (c *FakePools) DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error {
+	action := testing.NewDeleteCollectionAction(poolsResource, c.ns, listOpts)
 
 	_, err := c.Fake.Invokes(action, &v1alpha1.PoolList{})
 	return err
 }
 
 // Patch applies the patch and returns the patched pool.
-func (c *FakePools) Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.Pool, err error) {
+func (c *FakePools) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.Pool, err error) {
 	obj, err := c.Fake.
 		Invokes(testing.NewPatchSubresourceAction(poolsResource, c.ns, name, pt, data, subresources...), &v1alpha1.Pool{})
 
