@@ -1,24 +1,25 @@
 /*
-Copyright The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
+ * Tencent is pleased to support the open source community by making TKEStack available.
+ *
+ * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at
+ *
+ * https://opensource.org/licenses/Apache-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 // Code generated by client-gen. DO NOT EDIT.
 
 package v1alpha1
 
 import (
+	"context"
 	"time"
 
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -37,14 +38,14 @@ type FloatingIPsGetter interface {
 
 // FloatingIPInterface has methods to work with FloatingIP resources.
 type FloatingIPInterface interface {
-	Create(*v1alpha1.FloatingIP) (*v1alpha1.FloatingIP, error)
-	Update(*v1alpha1.FloatingIP) (*v1alpha1.FloatingIP, error)
-	Delete(name string, options *v1.DeleteOptions) error
-	DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error
-	Get(name string, options v1.GetOptions) (*v1alpha1.FloatingIP, error)
-	List(opts v1.ListOptions) (*v1alpha1.FloatingIPList, error)
-	Watch(opts v1.ListOptions) (watch.Interface, error)
-	Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.FloatingIP, err error)
+	Create(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.CreateOptions) (*v1alpha1.FloatingIP, error)
+	Update(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.UpdateOptions) (*v1alpha1.FloatingIP, error)
+	Delete(ctx context.Context, name string, opts v1.DeleteOptions) error
+	DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error
+	Get(ctx context.Context, name string, opts v1.GetOptions) (*v1alpha1.FloatingIP, error)
+	List(ctx context.Context, opts v1.ListOptions) (*v1alpha1.FloatingIPList, error)
+	Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error)
+	Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.FloatingIP, err error)
 	FloatingIPExpansion
 }
 
@@ -61,19 +62,19 @@ func newFloatingIPs(c *GalaxyV1alpha1Client) *floatingIPs {
 }
 
 // Get takes name of the floatingIP, and returns the corresponding floatingIP object, and an error if there is any.
-func (c *floatingIPs) Get(name string, options v1.GetOptions) (result *v1alpha1.FloatingIP, err error) {
+func (c *floatingIPs) Get(ctx context.Context, name string, options v1.GetOptions) (result *v1alpha1.FloatingIP, err error) {
 	result = &v1alpha1.FloatingIP{}
 	err = c.client.Get().
 		Resource("floatingips").
 		Name(name).
 		VersionedParams(&options, scheme.ParameterCodec).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // List takes label and field selectors, and returns the list of FloatingIPs that match those selectors.
-func (c *floatingIPs) List(opts v1.ListOptions) (result *v1alpha1.FloatingIPList, err error) {
+func (c *floatingIPs) List(ctx context.Context, opts v1.ListOptions) (result *v1alpha1.FloatingIPList, err error) {
 	var timeout time.Duration
 	if opts.TimeoutSeconds != nil {
 		timeout = time.Duration(*opts.TimeoutSeconds) * time.Second
@@ -83,13 +84,13 @@ func (c *floatingIPs) List(opts v1.ListOptions) (result *v1alpha1.FloatingIPList
 		Resource("floatingips").
 		VersionedParams(&opts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Watch returns a watch.Interface that watches the requested floatingIPs.
-func (c *floatingIPs) Watch(opts v1.ListOptions) (watch.Interface, error) {
+func (c *floatingIPs) Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error) {
 	var timeout time.Duration
 	if opts.TimeoutSeconds != nil {
 		timeout = time.Duration(*opts.TimeoutSeconds) * time.Second
@@ -99,66 +100,69 @@ func (c *floatingIPs) Watch(opts v1.ListOptions) (watch.Interface, error) {
 		Resource("floatingips").
 		VersionedParams(&opts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Watch()
+		Watch(ctx)
 }
 
 // Create takes the representation of a floatingIP and creates it.  Returns the server's representation of the floatingIP, and an error, if there is any.
-func (c *floatingIPs) Create(floatingIP *v1alpha1.FloatingIP) (result *v1alpha1.FloatingIP, err error) {
+func (c *floatingIPs) Create(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.CreateOptions) (result *v1alpha1.FloatingIP, err error) {
 	result = &v1alpha1.FloatingIP{}
 	err = c.client.Post().
 		Resource("floatingips").
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(floatingIP).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Update takes the representation of a floatingIP and updates it. Returns the server's representation of the floatingIP, and an error, if there is any.
-func (c *floatingIPs) Update(floatingIP *v1alpha1.FloatingIP) (result *v1alpha1.FloatingIP, err error) {
+func (c *floatingIPs) Update(ctx context.Context, floatingIP *v1alpha1.FloatingIP, opts v1.UpdateOptions) (result *v1alpha1.FloatingIP, err error) {
 	result = &v1alpha1.FloatingIP{}
 	err = c.client.Put().
 		Resource("floatingips").
 		Name(floatingIP.Name).
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(floatingIP).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Delete takes name of the floatingIP and deletes it. Returns an error if one occurs.
-func (c *floatingIPs) Delete(name string, options *v1.DeleteOptions) error {
+func (c *floatingIPs) Delete(ctx context.Context, name string, opts v1.DeleteOptions) error {
 	return c.client.Delete().
 		Resource("floatingips").
 		Name(name).
-		Body(options).
-		Do().
+		Body(&opts).
+		Do(ctx).
 		Error()
 }
 
 // DeleteCollection deletes a collection of objects.
-func (c *floatingIPs) DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error {
+func (c *floatingIPs) DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error {
 	var timeout time.Duration
-	if listOptions.TimeoutSeconds != nil {
-		timeout = time.Duration(*listOptions.TimeoutSeconds) * time.Second
+	if listOpts.TimeoutSeconds != nil {
+		timeout = time.Duration(*listOpts.TimeoutSeconds) * time.Second
 	}
 	return c.client.Delete().
 		Resource("floatingips").
-		VersionedParams(&listOptions, scheme.ParameterCodec).
+		VersionedParams(&listOpts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Body(options).
-		Do().
+		Body(&opts).
+		Do(ctx).
 		Error()
 }
 
 // Patch applies the patch and returns the patched floatingIP.
-func (c *floatingIPs) Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.FloatingIP, err error) {
+func (c *floatingIPs) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.FloatingIP, err error) {
 	result = &v1alpha1.FloatingIP{}
 	err = c.client.Patch(pt).
 		Resource("floatingips").
-		SubResource(subresources...).
 		Name(name).
+		SubResource(subresources...).
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(data).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
