@@ -1,24 +1,25 @@
 /*
-Copyright The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
+ * Tencent is pleased to support the open source community by making TKEStack available.
+ *
+ * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at
+ *
+ * https://opensource.org/licenses/Apache-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 // Code generated by client-gen. DO NOT EDIT.
 
 package v1alpha1
 
 import (
+	"context"
 	"time"
 
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -37,14 +38,14 @@ type PoolsGetter interface {
 
 // PoolInterface has methods to work with Pool resources.
 type PoolInterface interface {
-	Create(*v1alpha1.Pool) (*v1alpha1.Pool, error)
-	Update(*v1alpha1.Pool) (*v1alpha1.Pool, error)
-	Delete(name string, options *v1.DeleteOptions) error
-	DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error
-	Get(name string, options v1.GetOptions) (*v1alpha1.Pool, error)
-	List(opts v1.ListOptions) (*v1alpha1.PoolList, error)
-	Watch(opts v1.ListOptions) (watch.Interface, error)
-	Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.Pool, err error)
+	Create(ctx context.Context, pool *v1alpha1.Pool, opts v1.CreateOptions) (*v1alpha1.Pool, error)
+	Update(ctx context.Context, pool *v1alpha1.Pool, opts v1.UpdateOptions) (*v1alpha1.Pool, error)
+	Delete(ctx context.Context, name string, opts v1.DeleteOptions) error
+	DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error
+	Get(ctx context.Context, name string, opts v1.GetOptions) (*v1alpha1.Pool, error)
+	List(ctx context.Context, opts v1.ListOptions) (*v1alpha1.PoolList, error)
+	Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error)
+	Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.Pool, err error)
 	PoolExpansion
 }
 
@@ -63,20 +64,20 @@ func newPools(c *GalaxyV1alpha1Client, namespace string) *pools {
 }
 
 // Get takes name of the pool, and returns the corresponding pool object, and an error if there is any.
-func (c *pools) Get(name string, options v1.GetOptions) (result *v1alpha1.Pool, err error) {
+func (c *pools) Get(ctx context.Context, name string, options v1.GetOptions) (result *v1alpha1.Pool, err error) {
 	result = &v1alpha1.Pool{}
 	err = c.client.Get().
 		Namespace(c.ns).
 		Resource("pools").
 		Name(name).
 		VersionedParams(&options, scheme.ParameterCodec).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // List takes label and field selectors, and returns the list of Pools that match those selectors.
-func (c *pools) List(opts v1.ListOptions) (result *v1alpha1.PoolList, err error) {
+func (c *pools) List(ctx context.Context, opts v1.ListOptions) (result *v1alpha1.PoolList, err error) {
 	var timeout time.Duration
 	if opts.TimeoutSeconds != nil {
 		timeout = time.Duration(*opts.TimeoutSeconds) * time.Second
@@ -87,13 +88,13 @@ func (c *pools) List(opts v1.ListOptions) (result *v1alpha1.PoolList, err error)
 		Resource("pools").
 		VersionedParams(&opts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Watch returns a watch.Interface that watches the requested pools.
-func (c *pools) Watch(opts v1.ListOptions) (watch.Interface, error) {
+func (c *pools) Watch(ctx context.Context, opts v1.ListOptions) (watch.Interface, error) {
 	var timeout time.Duration
 	if opts.TimeoutSeconds != nil {
 		timeout = time.Duration(*opts.TimeoutSeconds) * time.Second
@@ -104,71 +105,74 @@ func (c *pools) Watch(opts v1.ListOptions) (watch.Interface, error) {
 		Resource("pools").
 		VersionedParams(&opts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Watch()
+		Watch(ctx)
 }
 
 // Create takes the representation of a pool and creates it.  Returns the server's representation of the pool, and an error, if there is any.
-func (c *pools) Create(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err error) {
+func (c *pools) Create(ctx context.Context, pool *v1alpha1.Pool, opts v1.CreateOptions) (result *v1alpha1.Pool, err error) {
 	result = &v1alpha1.Pool{}
 	err = c.client.Post().
 		Namespace(c.ns).
 		Resource("pools").
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(pool).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Update takes the representation of a pool and updates it. Returns the server's representation of the pool, and an error, if there is any.
-func (c *pools) Update(pool *v1alpha1.Pool) (result *v1alpha1.Pool, err error) {
+func (c *pools) Update(ctx context.Context, pool *v1alpha1.Pool, opts v1.UpdateOptions) (result *v1alpha1.Pool, err error) {
 	result = &v1alpha1.Pool{}
 	err = c.client.Put().
 		Namespace(c.ns).
 		Resource("pools").
 		Name(pool.Name).
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(pool).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
 
 // Delete takes name of the pool and deletes it. Returns an error if one occurs.
-func (c *pools) Delete(name string, options *v1.DeleteOptions) error {
+func (c *pools) Delete(ctx context.Context, name string, opts v1.DeleteOptions) error {
 	return c.client.Delete().
 		Namespace(c.ns).
 		Resource("pools").
 		Name(name).
-		Body(options).
-		Do().
+		Body(&opts).
+		Do(ctx).
 		Error()
 }
 
 // DeleteCollection deletes a collection of objects.
-func (c *pools) DeleteCollection(options *v1.DeleteOptions, listOptions v1.ListOptions) error {
+func (c *pools) DeleteCollection(ctx context.Context, opts v1.DeleteOptions, listOpts v1.ListOptions) error {
 	var timeout time.Duration
-	if listOptions.TimeoutSeconds != nil {
-		timeout = time.Duration(*listOptions.TimeoutSeconds) * time.Second
+	if listOpts.TimeoutSeconds != nil {
+		timeout = time.Duration(*listOpts.TimeoutSeconds) * time.Second
 	}
 	return c.client.Delete().
 		Namespace(c.ns).
 		Resource("pools").
-		VersionedParams(&listOptions, scheme.ParameterCodec).
+		VersionedParams(&listOpts, scheme.ParameterCodec).
 		Timeout(timeout).
-		Body(options).
-		Do().
+		Body(&opts).
+		Do(ctx).
 		Error()
 }
 
 // Patch applies the patch and returns the patched pool.
-func (c *pools) Patch(name string, pt types.PatchType, data []byte, subresources ...string) (result *v1alpha1.Pool, err error) {
+func (c *pools) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts v1.PatchOptions, subresources ...string) (result *v1alpha1.Pool, err error) {
 	result = &v1alpha1.Pool{}
 	err = c.client.Patch(pt).
 		Namespace(c.ns).
 		Resource("pools").
-		SubResource(subresources...).
 		Name(name).
+		SubResource(subresources...).
+		VersionedParams(&opts, scheme.ParameterCodec).
 		Body(data).
-		Do().
+		Do(ctx).
 		Into(result)
 	return
 }
