@@ -1,19 +1,19 @@
 /*
-Copyright The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
+ * Tencent is pleased to support the open source community by making TKEStack available.
+ *
+ * Copyright (C) 2012-2019 Tencent. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at
+ *
+ * https://opensource.org/licenses/Apache-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 // Code generated by lister-gen. DO NOT EDIT.
 
 package v1alpha1
@@ -26,8 +26,10 @@ import (
 )
 
 // PoolLister helps list Pools.
+// All objects returned here must be treated as read-only.
 type PoolLister interface {
 	// List lists all Pools in the indexer.
+	// Objects returned here must be treated as read-only.
 	List(selector labels.Selector) (ret []*v1alpha1.Pool, err error)
 	// Pools returns an object that can list and get Pools.
 	Pools(namespace string) PoolNamespaceLister
@@ -58,10 +60,13 @@ func (s *poolLister) Pools(namespace string) PoolNamespaceLister {
 }
 
 // PoolNamespaceLister helps list and get Pools.
+// All objects returned here must be treated as read-only.
 type PoolNamespaceLister interface {
 	// List lists all Pools in the indexer for a given namespace.
+	// Objects returned here must be treated as read-only.
 	List(selector labels.Selector) (ret []*v1alpha1.Pool, err error)
 	// Get retrieves the Pool from the indexer for a given namespace and name.
+	// Objects returned here must be treated as read-only.
 	Get(name string) (*v1alpha1.Pool, error)
 	PoolNamespaceListerExpansion
 }
