@@ -63,7 +63,7 @@ func newFakeCrdCache() (CrdCache, chan struct{}) {
 	dynamicClient := dynamic.NewSimpleDynamicClient(runtime.NewScheme(), objHasReplicas, objNotScalable)
 	extensionClient := extensionclient.NewSimpleClientset(FooCrd, NotScalableCrd)
 	extensionFactory := extensioninformer.NewSharedInformerFactory(extensionClient, 0)
-	extensionInformer := extensionFactory.Apiextensions().V1beta1().CustomResourceDefinitions()
+	extensionInformer := extensionFactory.Apiextensions().V1().CustomResourceDefinitions()
 	extensionInformer.Informer() // call Informer to actually create an informer
 	extensionFactory.Start(stop)
 	extensionFactory.WaitForCacheSync(stop)
